\name{initialize}
\docType{methods}
\alias{initialize}
\alias{initialize-methods}
\alias{initialize,char,char-method}
\title{Instantiate one Rvle object}

\description{
Instantiates one Rvle Object.
}

\usage{
 	new("Rvle",file = filename, pkg = pkgname)
}

\arguments{
 \item{filename}{the name of a vpz file.}
 \item{pkgname}{the name of a package that contains the vpz filename.}
}

\value{
	A Rvle object
}

