\name{is.rvle}
\docType{methods}
\alias{is.rvle}

\title{Check if an object is a VLE object.}

\description{
Check if an object is an VLE object. A VLE object is built using 
rvle_open function
}
\usage{
 	is.rvle(object)
}

\arguments{
 \item{object}{any type of object.}
}

\value{
 TRUE if object is a rvle object, FALSE otherwise. 
}
