\name{.restorebackup}
\docType{methods}
\alias{.restorebackup}
\alias{.restorebackup-methods}
\alias{.restorebackup,Rvle-method}
\title{Restores the backup settings.}

\description{
Restores the registered backup settings registered so far.
}

\usage{
 	.restorebackup(RvleObj)
}

\arguments{
 \item{RvleObj}{a Rvle object.}
}

\section{Note}{
	This is an internal function and should be used carefully.
}
