\name{run}
\docType{methods}
\alias{run}
\alias{run-methods}
\alias{run,Rvle,ANY-method}
\title{Run simulations of a vpz.}
\description{
	Runs simulations of a vpz. It has the same signature 
	as function 'setDefault'. See help(setDefault)
}
\usage{
 	run(RvleObj,...)
}
\arguments{
 \item{RvleObj}{a Rvle object}
 \item{...}{local settings for simulation. The current settings will be restored 
 after simulation. For more details on setting, type help(setDefault)}
}
