\name{Rvle Class}
\docType{class}
\alias{Rvle}
\alias{Rvle-class}

\title{Rvle S4 class}
\description{
The Rvle S4 class provides user-friendly functionnalities for manipulating
and simulating VLE models, it relies on the basic API rvle. See 'help(rvle)'

To get information on methods of the class, type 'help(methodname)' where 
methodname is one of the methods listed here.
}
\section{Methods}{
    results : show results of the last simulation process
    
    saveVpz : save a Vpz to a file
    
    show : show the current settings of Rvle object
    
    run : run simulations of VLE
    
    setDefault : set default settings to the Rvle object 
    
    config<- : set default experiment settings (Deprecated)
    
    config : show default experiment settings (Deprecated)
}
