
\name{rvle.manager_clear}
\docType{methods}
\alias{rvle.manager_clear}

\title{Reset simulation options.}

\description{
The VLE software provides the simulation of experiment plans. 
This functions resets the default 
simulation 
options.  
}

\usage{
 rvle.manager_clear(vleObj)
}

\arguments{
 \item{vleObj}{a VLE object, initialized with function rvle_open}
}

\seealso{
 rvle.manager_set_config,
 rvle.manager_get_config
}



