\name{rvle.packages_list}
\docType{methods}
\alias{rvle.packages_list}

\title{List the VLE package installed on the system.}

\description{
The VLE software relies on a package management system to add plugins, 
code of atomic models, etc.. This function lists the VLE packages installed on 
the system.
}

\usage{
 rvle.packages_list()
}

\value{
 the list of names of VLE packages installed on the system.  
}

\seealso{
 rvle.package_content
}

