
\name{rvle.plan_clear}
\docType{methods}
\alias{rvle.plan_clear}

\title{Reset experiment plan}

\description{
The VLE software provides the simulation of experiment plans. 
This functions clear the current experiment plans.  
}

\usage{
 rvle.plan_clear(vleObj)
}

\arguments{
 \item{vleObj}{a VLE object, initialized with function rvle_open}
}

\seealso{
 rvle.plan_input,
 rvle.plan_replicate,
 rvle.plan_set_config,
}



