
\name{rvle.plan_get}
\docType{methods}
\alias{rvle.plan_get}

\title{Get the current simulation plan.}

\description{
The VLE software provides the simulation of experiment plans. 
This function return the plan as defined by registerd inputs, outputs , etc..  
}

\usage{
 rvle.plan_get(vleObj)
}

\arguments{
 \item{vleObj}{a VLE object, initialized with function rvle_open}  
}

\seealso{
 rvle.plan_input,
 rvle.plan_propagate,
 rvle.plan_replicate
}



