
\name{rvle.plan_get_config}
\docType{methods}
\alias{rvle.plan_get_config}

\title{Get the configuration of the manager.}

\description{
The VLE software provides the simulation of experiment plans based on 
parallelization techniques. This function returns the configuration of the 
simulation manager (parallel option, nb slots, etc..) 
}

\usage{
 rvle.plan_get_config(vleObj)
}

\arguments{
 \item{vleObj}{a VLE object, initialized with function rvle_open}
}

\seealso{
 rvle.plan_set_config
}


