
\name{rvle.plan_reset}
\docType{methods}
\alias{rvle.plan_reset}

\title{Reset experiment plan and simulation options.}

\description{
The VLE software provides the simulation of experiment plans. 
This functions clear the current experiment plans and reset the default simulation 
options.  
}

\usage{
 rvle.plan_reset(vleObj)
}

\arguments{
 \item{vleObj}{a VLE object, initialized with function rvle_open}  
}

\seealso{
 rvle.plan_input,
 rvle.plan_replicate,
 rvle.plan_set_config,
}



