\name{rvle.save}
\docType{methods}
\alias{rvle.save}

\title{Save a VPZ with configured with the default values to a file}

\description{
A VLE model is defined into an xml file which type is VPZ.
This functions saves the VPZ withe teh default values set to a file.
}

\usage{
 rvle.save(vleObj, filename)
}

\arguments{
 \item{vleObj}{a VLE object, initialized with the function rvle_open},
 \item{filename}{relative path where is saved the VPZ model}
}

\seealso{
 rvle.open,
 rvle.show
}








