\name{rvle.set_log_level}
\docType{methods}
\alias{rvle.set_log_level}

\title{Set the level of logging during the simulation of the model}

\description{
Sets the level logging level between 0 and  7. If the level is >  4, then debug
messages are given.
}

\usage{
 rvle.set_log_level(vleObj, level)
}

\arguments{
 \item{vleObj}{a VLE object, initialized with function rvle_open}
 \item{level}{level of logging.}
}

\seealso{
 rvle.set_log_level
}


