\name{rvle.show}
\docType{methods}
\alias{rvle.show}

\title{Shows the content of a VPZ model}

\description{
A VLE model is defined into an xml file which type is VPZ.
This functions shows the content of a model.  
}

\usage{
 rvle.show(vleObj)
}

\arguments{
 \item{vleObj}{a VLE object, initialized with the function rvle_open}
}

\seealso{
 rvle.open,
 rvle.save
}








