\name{rvle.addBooleanCondition}
\alias{rvle.addBooleanCondition}
\title{Add a boolean to the specified condition port.}
\description{Add a boolean to the specified condition port.}
\usage{rvle.addBooleanCondition(self, condition, port, value)}
\arguments{
        \item{self}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{a boolean}
}
\keyword{models}
\keyword{optimize}

