\name{rvle.addCondition}
\docType{methods}
\alias{rvle.addCondition}
\title{Add a condition to a vpz}
\description{Add a condition to a vpz.
	Part of the basic API of rvle.
}
\usage{rvle.addCondition(rvleHandle, condition)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{condition}{string representation of the condition to add}
}
\keyword{models}
\keyword{optimize}

