\name{rvle.addObservablePort}
\docType{methods}
\alias{rvle.addObservablePort}
\title{Add a port to an observable of a VPZ.}
\description{
   Add a port to an observable of a VPZ. 
   Part of the basic API of rvle.
}
\usage{
 rvle.addObservablePort(rvleHandle, obsName, portName)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{obsName}{name of an observable of the vpz.}
 \item{portName}{name of the port to add to the observable.}
}

