\name{rvle.addPort}
\docType{methods}
\alias{rvle.addPort}
\title{Add a port to a condition of a vpz}
\description{Add a port to a condition of a vpz.
	Part of the basic API of rvle.
}
\usage{rvle.addPort(rvleHandle, condition, port)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the port to add}
}
\keyword{models}
\keyword{optimize}

