\name{rvle.addView}
\docType{methods}
\alias{rvle.addView}
\title{Add a view to a vpz}
\description{Add a view to a vpz.
	Part of the basic API of rvle.
}
\usage{rvle.addView(rvleHandle, view)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{view}{character specifying the view name to add}
}
\keyword{models}
\keyword{optimize}

