\name{rvle.attachCondition}
\docType{methods}
\alias{rvle.attachCondition}
\title{Attach a condition to an atomic model of a vpz}
\description{Attach a condition to an atomic model of a vpz.
	Part of the basic API of rvle.
}
\usage{rvle.attachCondition(rvleHandle, path_to_atomic, condition)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{path_to_atomic}{path to an atomic model}
        \item{condition}{string representation of the condition}
}
\keyword{models}
\keyword{optimize}

