\name{rvle.attachView}
\docType{methods}
\alias{rvle.attachView}
\title{Attach a view to an observable port of a VPZ.}
\description{Attach a view to an observable port of a VPZ.
	Part of the basic API of rvle.
}
\usage{rvle.attachView(rvleHandle, view, obsName, portName)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{view}{character specifying the view name}
        \item{obsName}{observable name}
       \item{portName}{name of the port of observable to be attached to view}
}
\keyword{models}
\keyword{optimize}

