\name{rvle.clearConditionPort}
\docType{methods}
\alias{rvle.clearConditionPort}
\title{Clear values from a condition port of VPZ.}
\description{
	Clear all values from a condition port of VPZ managed by 
	a rvle hanle. Part of the basic API of rvle.
	
}
\usage{
 rvle.clearConditionPort(rvleHandle, condition, port)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{condition}{name of the condition handled by rvleHandle.} 
 \item{port}{name of the port of the condition.}
}
