\name{rvle}
\alias{rvle.conditionPortListSize}
\title{Get the number of port for a specific condition.}
\description{Get the number of port for a specific condition.}
\usage{conditionPortListSize(f, condition)}
\arguments{
	\item{f}{handle of the rvle}
	\item{condition}{string representation of the condition}
}
\keyword{models}
\keyword{optimize}
