\name{rvle.detachCondition}
\docType{methods}
\alias{rvle.detachCondition}
\title{Detach a condition to an atomic model of a vpz}
\description{Detach a condition to an atomic model of a vpz.
	Part of the basic API of rvle.
}
\usage{rvle.detachCondition(rvleHandle, path_to_atomic, condition)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{path_to_atomic}{path to an atomic model}
        \item{condition}{string representation of the condition}
}
\keyword{models}
\keyword{optimize}

