\name{rvle.detachView}
\docType{methods}
\alias{rvle.detachView}
\title{Detach a view from an observable port of a VPZ.}
\description{Detach a view from an observable port of a VPZ.
    Part of the basic API of rvle.
}
\usage{rvle.detachView(rvleHandle, view, obsName, portName)}
\arguments{
       \item{rvleHandle}{a rvle handle.}
       \item{view}{character specifying the view name}
       \item{obsName}{observable name}
       \item{portName}{name of the port of observable to be detached from view}
}
\keyword{models}
\keyword{optimize}

