
\name{rvleExp.dateFromNum}
\docType{methods}
\alias{rvleExp.dateFromNum}

\title{Converts a julian day to a character representation of a date}

\description{
Utility function that converts a julian day to a charachater representation
of a date.
}

\usage{
 rvleExp.dateFromNum(dateNum, round)
}

\arguments{
 \item{dateNum}{a numeric representing a julian day eg. 2456659.5}
 \item{round}{if TRUE, the dateNum is rounded to match an exact date}
}

\value{
a character representing a date, eg : "2014-01-01"
}

\seealso{
 rvleExp.dateFromNum
}



