
\name{rvleExp.dateToNum}
\docType{methods}
\alias{rvleExp.dateToNum}

\title{Converts a character to a julian day}

\description{
Utility function that converts a string date as a julian day.
}

\usage{
 rvleExp.dateToNum(dateStr)
}

\arguments{
 \item{dateStr}{a character representing a date, eg : "2014-01-01"},
}

\value{
the numeric julian day, eg. : 2456659
}

\seealso{
 rvleExp.dateFromNum
}



