\name{rvle.getAttachedViewsSize}
\docType{methods}
\alias{rvle.getAttachedViewsSize}
\title{get the number of attached views to an observable port.}
\description{get the number of attached views to an observable port.
	Part of the basic API of rvle.
}
\usage{rvle.getAttachedViewsSize(rvleHandle, obsName, portName)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{obsName}{observable name}
       \item{portName}{name of the port of observable which attached views 
       have to be numbered.}
}
\keyword{models}
\keyword{optimize}

