\name{rvle.getBegin}
\docType{methods}
\alias{rvle.getBegin}
\title{Get the begin date of a VPZ.}
\description{
	Get the begin date of a VPZ managed by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.getBegin(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
	the begin date of the simulation
}
