\name{rvle.getConditionPortsSize}
\docType{methods}
\alias{rvle.getConditionPortsSize}
\title{Get the number of port of a condition.}
\description{
	Get the number of port of a condition
	of a VPZ managed by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.getConditionPortsSize(rvleHandle, condition)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{condition}{name of the condition}
}
\value{
 the number of ports of the condition 'condition'
}

