\name{rvle.getConditionsSize}
\docType{methods}
\alias{rvle.getConditionsSize}
\title{Gets the number of conditions in VPZ.}
\description{
	Gets the number of conditions in VPZ
	managed by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.getConditionsSize(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
 the number of conditions of the VPZ managed by 'rvleHandle'.
}
