\name{rvle.getConfigView}
\docType{methods}
\alias{rvle.getConfigView}
\title{Get the configuration of view of a VPZ.}
\description{
	Get the configuration of view of a VPZ..
	Part of the basic API of rvle.
}
\usage{
 rvle.getConfigView(rvleHandle, view)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{view}{the name of the view.}
}
\value{
 the configuration type of the view 'view'
 into the VPZ represented by the rvle handle 'rvleHandle'
}
