\name{rvle.getDuration}
\docType{methods}
\alias{rvle.getDuration}
\title{Get the duration of the experiment.}
\description{
	Gets the duration of the experiment by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.getDuration(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
 the duraton of the experiment handled by 'rvleHandle'
}
