\name{rvle.getObservablePortsSize}
\docType{methods}
\alias{rvle.getObservablePortsSize}
\title{Get the number of ports of an observables of a VPZ.}
\description{
	Get the number of ports of an observables of a VPZ
	managed by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.getObservablePortsSize(rvleHandle, obsName)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{obsName}{the name of an observable}
}
\value{
 the number of ports of an observable of the VPZ managed by 'rvleHandle'.
}
