\name{rvle.getObservablesSize}
\docType{methods}
\alias{rvle.getObservablesSize}
\title{Get the number of observables of a VPZ.}
\description{
	Get the number of observables of a VPZ 
	managed by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.getObservablesSize(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
 the number of observables of the VPZ managed by 'rvleHandle'.
}
