\name{rvle.getOutputPlugin}
\docType{methods}
\alias{rvle.getOutputPlugin}
\title{Get the output plugin name of a VPZ.}
\description{
	Get the output plugin name of a VPZ, represented by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.getOutputPlugin(rvleHandle, view)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{view}{the name of the view.}
}
\value{
 the name of the output plugin attached to view 'view' for 
 into the VPZ represented by the rvle handle 'rvleHandle'
}
