\name{rvle.getSeed}
\docType{methods}
\alias{rvle.getSeed}
\title{Get the seed value of a VPZ.}
\description{
	Get the seed value of a VPZ represented by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.getSeed(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
 the seed value of the of the VPZ represented by the rvle handle 'rvleHandle'
}
