\name{rvle.getViewsSize}
\docType{methods}
\alias{rvle.getViewsSize}
\title{Get the number of views in a VPZ.}
\description{
	Gets the number of views in VPZ
	managed by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.getViewsSize(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
 the number of views of the VPZ managed by 'rvleHandle'.
}
