\name{rvle.listAttachedViews}
\docType{methods}
\alias{rvle.listAttachedViews}
\title{Attach a view to an observable port of a VPZ.}
\description{Attach a view to an observable port of a VPZ.
	Part of the basic API of rvle.
}
\usage{rvle.listAttachedViews(rvleHandle, obsName, portName)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{obsName}{observable name}
       \item{portName}{name of the port of observable which attached views are 
       required}
}
\keyword{models}
\keyword{optimize}

