\name{rvle.listConditions}
\docType{methods}
\alias{rvle.listConditions}
\title{Get the list of condition names of the VPZ.}
\description{
	Get the list of condition names of the VPZ
	managed by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.listConditions(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
 the list of condition names of the VPZ managed by 'rvleHandle'.
}
