\name{rvle.listObservablePorts}
\docType{methods}
\alias{rvle.listObservablePorts}
\title{Get the list of ports of an observable of a VPZ.}
\description{
	Get the list of ports of an observable of a VPZ.
	managed by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.listObservablePorts(rvleHandle, obsName)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{ obsName}{the name of the observable.}
}
\value{
 the list of ports of the observable obsName of the VPZ managed by 'rvleHandle'.
}

