\name{rvle.listObservables}
\docType{methods}
\alias{rvle.listObservables}
\title{Get the list of observables of the VPZ.}
\description{
	Get the list of observables of the VPZ
	managed by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.listObservables(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
 the list of observables of the VPZ managed by 'rvleHandle'.
}

