\name{rvle.listPackages}
\docType{methods}
\alias{rvle.listPackages}
\title{List the vle packages available for simulation.}
\description{List the vle packages available for simulation.
    Part of the basic API of rvle.
}
\usage{rvle.listPackages(justprint=TRUE)}
\arguments{
     \item{justprint}{if TRUE, a print is performed, otherwise
                      the vector of string is returned}
}
\value{
 the list of vle packages that have been installed if justprint=FALSE
}
\keyword{models}
\keyword{optimize}

