\name{rvle.listViews}
\docType{methods}
\alias{rvle.listViews}
\title{Get the list of view names of the VPZ.}
\description{
	Get the list of view names of the VPZ
	managed by a rvle handle.
	Part of the basic API of rvle.
}
\usage{
 rvle.listViews(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
 the list of view names of the VPZ managed by 'rvleHandle'.
}

