\name{rvle.packageContent}
\docType{methods}
\alias{rvle.packageContent}
\title{Gives the content of a vle package (vpz, dynamics, etc..)}
\description{Gives the content of a vle package (vpz, dynamics, etc..)
    Part of the basic API of rvle.
}
\usage{rvle.packageContent(pkgname, justprint=TRUE)}
\arguments{
     \item{pkgname}{the name of a vle package installed into the VLE_HOME}
     \item{justprint}{if TRUE, a print is performed, otherwise
                      the vector of string is returned}
}
\value{
  the content of a vle package if justprint=FALSE
}
\keyword{models}
\keyword{optimize}

