\name{rvle.removeCondition}
\docType{methods}
\alias{rvle.removeCondition}
\title{Remove a condition of a vpz}
\description{Remove a condition of a vpz.
	Part of the basic API of rvle.
}
\usage{rvle.removeCondition(rvleHandle, condition)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{condition}{string representation of the condition}
}
\keyword{models}
\keyword{optimize}

