\name{rvle.removeObservablePort}
\docType{methods}
\alias{rvle.removeObservablePort}
\title{Remove a port from an observable of a VPZ.}
\description{
   Remove a port from an observable of a VPZ.
   Part of the basic API of rvle.
}

\usage{
 rvle.removeObservablePort(rvleHandle, obsName, portName)
}

\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{obsName}{name of an observable of the vpz.}
 \item{portName}{name of the port to remove to the observable.}
}

