\name{rvle.removePort}
\docType{methods}
\alias{rvle.removePort}
\title{Remove a port to a condition of a vpz}
\description{Remove a port to a condition of a vpz.
	Part of the basic API of rvle.
}
\usage{rvle.removePort(rvleHandle, condition, port)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the port to remove}
}
\keyword{models}
\keyword{optimize}

