\name{rvle.removeView}
\docType{methods}
\alias{rvle.removeView}
\title{Remove a view from a vpz}
\description{Remove a view from a vpz.
	Part of the basic API of rvle.
}
\usage{rvle.removeView(rvleHandle, view)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{view}{character specifying the view name to remove}
}
\keyword{models}
\keyword{optimize}

