\name{rvle.run}
\docType{methods}
\alias{rvle.run}
\title{Use the handle of rvle to launch a simulation.}
\description{
	Use the handle of rvle to launch a simulation.
	Part of the basic API of rvle.
}
\usage{
 rvle.run(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
Returns a list of dataframes.
}
