\name{rvle.runManagerMatrix}
\docType{methods}
\alias{rvle.runManagerMatrix}
\title{Use an experimental frame and launch simulation.
}
\description{
	Use an experimental frame and launch simulation.
	Part of the basic API of rvle.
}
\usage{
 rvle.runManagerMatrix(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
Returns a matrix of list of matrices of reals.
}


