\name{rvle.runMatrix}
\docType{methods}
\alias{rvle.runMatrix}
\title{Use the handle of rvle to launch a simulation.}
\description{
	Use the handle of rvle to launch a simulation.
	Part of the basic API of rvle.
}
\usage{
 rvle.runMatrix(rvleHandle)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
}
\value{
Returns a list of matrices of reals.
}


