\name{rvle.runPoly}
\docType{methods}
\alias{rvle.runPoly}
\title{Use the handle of rvle to launch a simulation.}
\description{
	Use the handle of rvle to launch a simulation.
	It can be used if returned values of simulations are complex data 
    (list, vector, matrices ...)
	Part of the basic API of rvle.
}
\usage{
 rvle.runPoly(rvleHandle, withSpawn = 1)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{withSpawn}{advanced option for simulation.}
}
\value{
Returns a list of 2D list.
}


