\name{rvle.save}
\docType{methods}
\alias{rvle.save}
\title{Write a vpz file under the specified filename.}
\description{
	Write a vpz file under the specified filename.
	Part of the basic API of rvle.
}
\usage{
 rvle.save(rvleHandle, file)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{file}{string representation of the filename where store vpz file}
}


