\name{rvle.setBegin}
\docType{methods}
\alias{rvle.setBegin}
\title{Set the begin of the experiment.}
\description{
	Set the begin of the experiment.
	Part of the basic API of rvle.
}
\usage{
 rvle.setBegin(rvleHandle, value)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{value}{the begin date of simulation.}
}

