\name{rvle.setBooleanCondition}
\alias{rvle.setBooleanCondition}
\title{Set a boolean to the specified condition port.}
\description{Set a boolean to the specified condition port.}
\usage{rvle.setBooleanCondition(self, condition, port, value)}
\arguments{
        \item{self}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{an boolean}
}
\keyword{models}
\keyword{optimize}

