\name{rvle.setConfigView}
\docType{methods}
\alias{rvle.setConfigView}
\title{Sets the configuration of view of a VPZ.}
\description{
	Sets the configuration of view of a VPZ..
	Part of the basic API of rvle.
}
\usage{
 rvle.setConfigView(rvleHandle, view, config)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{view}{the name of the view.}
 \item{config}{the configuration of view: either 'finish' or 
                  'timed,ts' where ts is the time step}
}

