\name{rvle.setDuration}
\docType{methods}
\alias{rvle.setDuration}
\title{Set the duration of the experiment.}
\description{
	Set the begin of the experiment.
	Part of the basic API of rvle.
}
\usage{
 rvle.setDuration(rvleHandle, value)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{value}{a real to set the duration}
}