\name{rvle.setIntegerCondition}
\docType{methods}
\alias{rvle.setIntegerCondition}
\title{Set a integer to the specified condition port.}
\description{
	Set a integer to the specified condition port.
	Part of the basic API of rvle.
	This function is DEPRECATED.  
}
\usage{
 rvle.setIntegerCondition(rvleHandle, condition, port, value)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{condition}{string representation of the condition}
 \item{port}{string representation of the condition's port}
 \item{value}{an integer}
}


