\name{rvle.setLinearCombination}
\docType{methods}

\alias{rvle.setLinearCombination}

\title{Assign a linear experimental frame with specific seed to build seeds of
simulations and the number of repliquas.
}
\description{Assign a linear experimental frame with specific seed to build
seeds of simulations and the number of repliquas.
Part of the basic API.
}
\usage{rvle.setLinearCombination(rvleHandle, seed, repliquas)}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{seed}{an integer to set the seed of random seed}
 \item{repliquas}{an integer to set the number of repliquas}
}

