\name{rvle.setOutputPlugin}
\docType{methods}
\alias{rvle.setOutputPlugin}
\title{Set the plugin output of a view.}
\description{Set the plugin output of a view.
	Part of the basic API of rvle.}
\usage{rvle.setOutputPlugin(rvleHandle, viewname, pluginname)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{viewname}{the name of the view.}
        \item{pluginname}{the name of the plugin}
}
\keyword{models}
\keyword{optimize}
