\name{rvle.setRealCondition}
\docType{methods}
\alias{rvle.setRealCondition}
\title{Set a real to the specified condition port.}
\description{Set a real to the specified condition port.
	Part of the basic API of rvle.}
\usage{rvle.setRealCondition(rvleHandle, condition, port, value)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{a real}
}
\keyword{models}
\keyword{optimize}

