\name{rvle.setSeed}
\alias{rvle.setSeed}
\title{Set the seed of the experiment.}
\description{Set the seed of the experiment.}
\usage{rvle.setSeed(self, value)}
\arguments{
        \item{self}{handle of the rvle}
        \item{value}{an integer to set the seed}
}
\keyword{models}
\keyword{optimize}
