\name{rvle.setStringCondition}
\docType{methods}
\alias{rvle.setStringCondition}
\title{Set a string to the specified condition port.}
\description{Set a string to the specified condition port.
	Part of the basic API of rvle.
	This function is DEPRECATED.  
}
\usage{rvle.setStringCondition(rvleHandle, condition, port, value)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{an string}
}
\keyword{models}
\keyword{optimize}

