\name{rvle.setTotalCombination}
\docType{methods}
\alias{rvle.setTotalCombination}
\title{Assign a total experimental frame with specific seed to build seeds of
    simulations and the number of repliquas.}
\description{Assign a total experimental frame with specific seed to build
    seeds of simulations and the number of repliquas.
	Part of the basic API of rvle.}
\usage{rvle.setTotalCombination(rvleHandle, seed, repliquas)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{seed}{an integer to set the seed of random seed}
        \item{repliquas}{an integer to set the number of repliquas}
}
\keyword{models}
\keyword{optimize}
