\name{rvle.setTupleCondition}
\alias{rvle.setTupleCondition}
\title{Set a tuple (list of reals) to the specified condition port.}
\description{Set a list of reals to the specified condition port.}
\usage{rvle.setTupleCondition(self, condition, port, value)}
\arguments{
        \item{self}{handle of the rvle}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{a list of reals}
}
\keyword{models}
\keyword{optimize}

