\name{rvle.setValueCondition}
\docType{methods}
\alias{rvle.setValueCondition}
\title{Set a R value to the specified condition port.}
\description{
	Set a R value to the specified condition port.
	Part of the basic API of rvle.  
}
\usage{
 rvle.setValueCondition(rvleHandle, condition, port, value)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{condition}{string representation of the condition}
 \item{port}{string representation of the condition's port}
 \item{value}{a R value (numeric, list, dataframe, logical, ...)}
}



