\name{saveVpz}
\docType{methods}
\alias{saveVpz}
\alias{saveVpz-methods}
\alias{saveVpz,Rvle,char-method}

\title{Save the VPZ into a file.}

\description{
Save the VPZ into a file.
}

\usage{
 saveVpz(RvleObj, file)
}

\arguments{
 \item{RvleObj}{a Rvle object.}
 \item{file}{the file name where to save the VPZ.}
}
