\name{is.rvle}
\docType{methods}
\alias{is.rvle}
\title{Check if an object is an rvle handle.}

\description{
Check if an object is an rvle handle.
Part of the basic API.
}
\usage{
 	is.rvle(object)
}

\arguments{
 \item{object}{any type of object.}
}
\value{
	TRUE if object is a rvle handle, FALSE otherwise. 
}
\note{
	This function is part of the simple API of rvle, consider using 
	the integrated API based on S4 classes. help(rvle)
	will provide more detailed informations.
}
