\name{rvle.runManager}
\docType{methods}
\alias{rvle.runManager}
\title{Use an experimental frame and launch simulation.}
\description{
	Use an experimental frame and launch simulation.
	Part of the basic API of rvle.
}
\usage{
 rvle.runManager(rvleHandle, withSpawn = 1)
}
\arguments{
 \item{rvleHandle}{a rvle handle.}
 \item{withSpawn}{advanced option for simulation.}
}
\value{
Returns a matrix of list of dataframes.
}

