\name{rvle.runManager}
\alias{rvle.runManager}
\title{Use an experimental frame and launch simulation.}
\description{Use an experimental frame from the rvle handle and launch
simulation for each instance of the experimental frame. Returns a matrix
of list of reals matrix.}
\usage{rvle.runManager(self, commonSeed = TRUE)}
\arguments{
	\item{self}{handle of rvle}
	\item{commonSeed}{if true, the same seed is used for the simulation
 of different combinations of one replica}
}
\keyword{models}
\keyword{optimize}

