\name{rvle.save}
\alias{rvle.save}
\title{Write a vpz file under the specified filename.}
\description{Write a vpz file under the specified filename.}
\usage{rvle.save(self, file)}
\arguments{
    \item{self}{handle of the rvle}
    \item{file}{string representation of the filename where store vpz file}
}
\keyword{models}
\keyword{optimize}

