\name{rvle.setSeed}
\alias{rvle.setSeed}
\title{Set the seed of the experiment.}
\description{Set the seed of the experiment.
	Part of the basic API of rvle.}
\usage{rvle.setSeed(rvleHandle, value)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{value}{an integer to set the seed}
}
\keyword{models}
\keyword{optimize}
