\name{rvle.setTupleCondition}
\docType{methods}
\alias{rvle.setTupleCondition}
\title{Set a tuple (list of reals) to the specified condition port.}
\description{Set a list of reals to the specified condition port.
	Part of the basic API of rvle.}
\usage{rvle.setTupleCondition(rvleHandle, condition, port, value)}
\arguments{
        \item{rvleHandle}{a rvle handle.}
        \item{condition}{string representation of the condition}
        \item{port}{string representation of the condition's port}
        \item{value}{a list of reals}
}
\keyword{models}
\keyword{optimize}

